requirements.txt
requests==2.22.0
pandas==0.23.1
psycopg2-binary==2.7.5
SQLAlchemy==1.3.13